requirements.txt
pandas==2.0.3
requests==2.31.0
google-cloud-bigquery==3.11.4
python-dotenv==1.0.0
